+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = audiostream2py
-version = 0.1.11
+version = 0.2.0
 url = https://github.com/i2mint/audiostream2py
 platforms = any
 description_file = README.md
